+++ app/Modules/Attachments/Views/attachments/_modal_attach_files.blade.php
@@ -65,7 +65,7 @@
             <div class="modal-content">
                 <div class="modal-header">
                     <h4 class="modal-title">@lang('bt.attach_files')</h4>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-hidden="true"></button>
+                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-hidden="false"></button>
                 </div>
                 <div class="modal-body">
                     <p class="fw-bold">@lang('bt.attach_files')</p>