+++ app/Modules/CompanyProfiles/Views/company_profiles/_modal_lookup.blade.php
@@ -5,7 +5,7 @@
         <div class="modal-content">
             <div class="modal-header">
                 <h4 class="modal-title">@lang('bt.change_company_profile')</h4>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-hidden="true"></button>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-hidden="false"></button>
             </div>
             <div class="modal-body">
                 <div id="modal-status-placeholder"></div>