+++ app/Modules/MailQueue/Views/mail_log/_modal_content.blade.php
@@ -10,7 +10,7 @@
         <div class="modal-content">
             <div class="modal-header">
                 <h4 class="modal-title">{{ $mail->subject }}</h4>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-hidden="true"></button>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-hidden="false"></button>
             </div>
             <div class="modal-body" style="overflow-y: auto;">