+++ app/Modules/Payments/Views/payments/_js_mail.blade.php
@@ -11,7 +11,6 @@
 
         const modal = bsModal('modal-mail-payment')
         modal.show()
-        modaleL = document.getElementById('modal-mail-payment')
         modaleL.addEventListener('shown.bs.modal', function () {
             new TomSelect('#to', tsconfig);
             new TomSelect('#cc', tsconfig);